+++ src/routers/usersRoutes/index.ts
@@ -1,0 +1,3 @@
+import usersCreate from "./usersCreate";
+
+export { usersCreate };
